// src/app/models/transaction.ts
import { TransactionState } from "./transaction-state";

export interface ITransaction {
    uniqueId: string;
    title: string;
    buyerId: string;
    sellerId: string;
    listingId: string;
    creationDate: Date;
    completionDate: Date;
    buyerAccepted: boolean;
    sellerAccepted: boolean;
    buyerCompleted: boolean;
    sellerCompleted: boolean;
    state: TransactionState;
}